+++ mobile/android/android-components/components/support/ktx/src/test/java/mozilla/components/support/ktx/android/content/ContextTest.kt
@@ -10,23 +10,32 @@ import android.content.Context
 import android.content.Intent
 import android.content.Intent.FLAG_ACTIVITY_NEW_TASK
 import android.content.pm.PackageManager.PERMISSION_GRANTED
+import androidx.test.core.app.ApplicationProvider
 import mozilla.components.support.test.argumentCaptor
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
+import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mockito.spy
 import org.mockito.Mockito.verify
 import org.robolectric.RobolectricTestRunner
-import org.robolectric.RuntimeEnvironment
-import org.robolectric.Shadows
+import org.robolectric.shadows.ShadowApplication
+import org.robolectric.shadows.ShadowProcess
 
 @RunWith(RobolectricTestRunner::class)
 class ContextTest {
+    private val context: Context
+        get() = ApplicationProvider.getApplicationContext()
+
+    @Before
+    fun setup() {
+        isMainProcess = null
+    }
+
     @Test
     fun `systemService() returns same service as getSystemService()`() {
-        val context = RuntimeEnvironment.application
-
         assertEquals(
             context.getSystemService(Context.INPUT_METHOD_SERVICE),
             context.systemService(Context.INPUT_METHOD_SERVICE))
@@ -42,7 +51,6 @@ class ContextTest {
 
     @Test
     fun `isOSOnLowMemory() should return the same as getMemoryInfo() lowMemory`() {
-        val context = RuntimeEnvironment.application
         val extensionFunctionResult = context.isOSOnLowMemory()
 
         val activityManager = context.systemService<ActivityManager>(Context.ACTIVITY_SERVICE)
@@ -56,8 +64,7 @@ class ContextTest {
 
     @Test
     fun `isPermissionGranted() returns same service as checkSelfPermission()`() {
-        val context = RuntimeEnvironment.application
-        val application = Shadows.shadowOf(context)
+        val application = ShadowApplication()
 
         assertEquals(
                 context.isPermissionGranted(WRITE_EXTERNAL_STORAGE),
@@ -72,7 +79,7 @@ class ContextTest {
 
     @Test
     fun `share invokes startActivity`() {
-        val context = spy(RuntimeEnvironment.application)
+        val context = spy(context)
         val argCaptor = argumentCaptor<Intent>()
 
         val result = context.share("https://mozilla.org")
@@ -81,5 +88,39 @@ class ContextTest {
 
         assertTrue(result)
         assertEquals(FLAG_ACTIVITY_NEW_TASK, argCaptor.value.flags)
+    }
+
+    @Test
+    fun `isMainProcess must only return true if we are in the main process`() {
+        val myPid = Int.MAX_VALUE
+
+        assertTrue(context.isMainProcess())
+
+        ShadowProcess.setPid(myPid)
+        isMainProcess = null
+
+        assertFalse(context.isMainProcess())
+    }
+
+    @Test
+    fun `runOnlyInMainProcess must only run if we are in the main process`() {
+        val myPid = Int.MAX_VALUE
+        var wasExecuted = false
+
+        context.runOnlyInMainProcess {
+            wasExecuted = true
+        }
+
+        assertTrue(wasExecuted)
+
+        wasExecuted = false
+        ShadowProcess.setPid(myPid)
+        isMainProcess = false
+
+        context.runOnlyInMainProcess {
+            wasExecuted = true
+        }
+
+        assertFalse(wasExecuted)
     }
 }